# UserSelect: Speed up group and distribution action registering