--- a/timApp/plugin/userselect/utils.py
+++ b/timApp/plugin/userselect/utils.py
@@ -1,0 +1,4 @@
+from datetime import timedelta
+
+# It can be useful to offset the time a little to ensure any checks for expired memberships can pass
+group_expired_offset = timedelta(seconds=1)