Tämä PR lisää UserSelect-komponenttiin `useActionQueues`-attribuutin, joka puskuroi ryhmä- ja oikeuslevitystoimintojen ajamista. Lisäksi pari pientä muutosta:

* Celerystä tulevat lokit merkitään `[CELERY]`-tagilla
* Lisätty `keydefaultdict`, joka toimii kuten `defaultdict`, mutta `default_factory` saa parametriksi avaimen, jolla alkio haettiin

Olennaisin puskuroinnin käyttötarkoitus ovat valtavat tentit, jossa tenttijöitä kirjataan sisään saliryhmiin UserSelectilla. Näissä tenteissa UserSelect on usein konfiguroitu seuraavasti:

```yaml
    changeGroup:
      changeTo: %%saliryhma%%
      allGroups:
        - sali1
        - sali2
        - sali3
      verify: true
```

Kun tenttijä "piipataan" sisään UserSelectilla, yllä oleva esimerkki tomii seuraavasti:

* Tarkistetaan, kuuluuko tenttijä johonkin listattuun saliryhmään. Jos on, laitetaan asiasta varoitus.
* Lisätään tenttijä saliin `%%saliryhma%%`
* Poistetaan tenttijä muista salirymistä

Massatenteissa (esim. valintakokeissa) on usein kymmeniä valvojia kirjaamassa monta tenttijää sisään. Koska nyt muutokset ajetaan suoraan tietokantaan, joudutaan synkronoimaan operaatiot lukituksella. Viime vuonna havaittiin, että tällainen suora tietokantaan ajo + synkronointi on pullonkaula massatenteissa: jossain tapauksissa valvoja on joutunut odottamaan jopa 10 sekuntia yhden tenttijän kirjaamiseksi. Toisen pullonkaulan aiheuttaa oikeuksien levitys ja siihen liittyvä lukitus.

Tämä PR lisää UserSelectiin käsittelyjonon, joka puretaan taustalla Celery-workerissa. Käsittelyjonon ansiosta valvojan ei tarvitse odottaa sisäänkirjauksen viimeistelyä, vaan se hoituu taustalla. Täten valvojan kokema sisäänkirjaus on nopeampi, ja käsittelyoperaatiot puskuroidaan yhdeksi operaatioksi vähentäen tietokantakutsuja.

## Nopeustestausta

### Asetukset

Ajoin nopeustestit lokaalisti, koska niiden toistaminen palvelimella on hieman hankalaa. Testien tarkoitus on vain verrata kahta tapaa, joten riittää lokaali setuppi.
Testausympäristö:

* CPU: 12 ydintä
* SSD-levy
* TIM-palvelimella 6 workeria
* Testipalvelimelle tehty 5000 testitunnusta tunnuksella `mass{1..5000}`. Tunnukset lisättiin ryhmään `userselectgroup0`
* Testipalvelimelle tehty myös ryhmät `userselectgroup{1..10}`
* Kuormitus ajettu admin-käyttäjällä

<details><summary>Testausskripti</summary>
<p>

Skripti on pastettu sellaisenaan ja vaatii muokkausta. Pohjana on käytetty [tim-api](https://gitlab.jyu.fi/dezhidki/tim-api-py)-kirjastoa.

```py
import math
import datetime
import multiprocessing

domain = "http://localhost"

def apply_user(num: int) -> float:
    import os

    os.environ["TIM_DOMAIN"] = domain
    os.environ["TIM_USERNAME"] = "dz"
    os.environ["TIM_PASSWORD"] = "dz"

    from tim import get_session

    session = get_session()

    now = datetime.datetime.now()

    session.post(f"{domain}/userSelect/apply", json={
        "par": {
            "doc_id": 158,
            "par_id": "ntk7H4WsuOCg"
        },
        "username": f"mass{num}"
    })

    time_diff = datetime.datetime.now() - now
    time_diff_ms = time_diff.total_seconds() * 1000
    return int(time_diff_ms)


MIN = 1
MAX = 1000
BATCH_SIZE = 12


if __name__ == "__main__":
    with multiprocessing.Pool(processes=BATCH_SIZE) as pool:
        now = datetime.datetime.now()
        results = pool.map(apply_user, range(MIN, MAX + 1))
        time_diff = datetime.datetime.now() - now
        time_diff_ms = time_diff.total_seconds() * 1000
        print(results)
        # print standard deviation and standard error of mean
        mean = sum(results) / len(results)
        std_dev_sample = math.sqrt(sum([(x - mean) ** 2 for x in results]) / (len(results) - 1))
        se = std_dev_sample / math.sqrt(len(results))
        print("total time:", time_diff_ms)
        print("req/sec:", len(results) / (time_diff_ms / 1000))
        print("mean:", mean)
        print(f"min: {min(results)}, max: {max(results)}")
        print(f"std dev: {std_dev_sample}")
        print(f"se: {se}")
        print(f"n over 1000: {sum(1 for x in results if x > 1000)}")

        import matplotlib.pyplot as plt
        
        # histogram with bins between 100 ms
        plt.hist(results, bins=[i for i in range(0, max(results), 100)])
        # set grid lines for x and y
        plt.grid(True)
        # set x ticks between 100 starting at 0
        plt.xticks(range(0, max(results), 500))
        # Set x axis title to "time (ms)" and y axis title to ("frequency")
        plt.xlabel("time (ms)")
        plt.ylabel("frequency")
        plt.show()
```

</p>
</details> 

<details><summary>Testausdokumentin markup</summary>
<p>

````md
``` {#user_select plugin="userSelect" id="ntk7H4WsuOCg"}
groups:
    - userselectgroup0
useActionQueues: true
actions:
    changeGroup:
      changeTo: userselectgroup2
      allGroups:
        - userselectgroup1
        - userselectgroup2
        - userselectgroup3
        - userselectgroup4
        - userselectgroup5
        - userselectgroup6
        - userselectgroup7
        - userselectgroup8
        - userselectgroup9
        - userselectgroup10
      verify: false

    distributeRight:
      - operation: confirm
        target: ["leikki"]

    setValue:
      - taskId: 158.salimeni
        value: "testvalue1"
```

#- {defaultplugin="textfield" readonly="view" .fieldCell id="24TLdKuuaitb"}
{#salimeni #}
````

</p>
</details> 

<details><summary>`timApp/tim_files/leikki.rights.initial`</summary>
<p>

```
{"require_confirm":true,"accessible_from":"2023-05-06T15:00:00+00:00","accessible_to":"2023-05-06T18:00:00+00:00","duration":null,"duration_from":null,"duration_to":null}
{"type":"changetime","email":"mass1@test.org","secs":3600,"timestamp":"2023-05-30T15:00:00+00:00"}
{"type":"changestarttimegroup","group":"userselectgroup1","starttime":"2023-05-06T15:11:00+00:00","timestamp":"2023-05-30T15:00:00+00:00"}
{"type":"changestarttimegroup","group":"userselectgroup2","starttime":"2023-05-06T15:12:00+00:00","timestamp":"2023-05-30T15:00:00+00:00"}
{"type":"changestarttimegroup","group":"userselectgroup3","starttime":"2023-05-06T15:13:00+00:00","timestamp":"2023-05-30T15:00:00+00:00"}
```

</p>
</details> 

### Tulokset

**`useActionQueues: false`**

Tämä edustaa nykyistä tilannetta masterissa.

```
total time: 82961.63 ms
req/sec: 12.05 ms
mean: 947.76 ms
min: 74 ms, max: 8933 ms
std dev: 1171.84 ms
se: 37.05 ms
n over 1000 ms: 285
```

Histogrammi kaikkien kutsujen kestosta:
![plot_userselect_synchronized_12batch_6workers](https://user-images.githubusercontent.com/5202606/236645745-ba0b24db-d8dc-45b7-9eb3-8d5030bf4d41.png)

Huomioita:

* 285 kutsun kohdalla jouduttu odottamaan yli 1 sekunti piippauksen loppuun
* Suurin osa kuitenkin silti nopeat, monet kutstut alle 100ms
* Yhteensä 1000 tenttijän prosessointiin meni 1 min 23 sek

**`useActionQueues: true`**

Tämä vastaa tämän PR:n muutoksia

```
total time: 30757.74 ms
req/sec: 32.51 ms
mean: 343.46 ms
min: 99 ms, max: 841 ms
std dev: 104.53 ms
se: 3.30 ms
n over 1000: 0
```

Histogrammi kaikkien kutsujen kestosta:
![plot_userselect_queried_12batch_6workers](https://user-images.githubusercontent.com/5202606/236645857-61c49143-a4b6-48db-9437-65e66dc5cbe5.png)

Huomioita:

* Noin 2,5 -kertainen prosessointivolyymi (req/sec)
* Mikään piippaus ei kestänyt yli 1 sekunnnin
* **Mutta:** koko jonon käsittelyyn meni noin 1 min 30 sek. Toisin sanoin piippauksen jälkeen meni noin 1,5 min kunnes tenttijä pääsi saliryhmään. Tässä toki pitää ottaa huomioon, että samaan aikaan yritettiin ajaa 1000 tenttijää niin nopeasti kuin pystyy, mikä kasvatti käsittelyjonon liian nopeasti suureksi. Oikeassa tentissä tuskin tällaiseen tilanteeseen päädytään. 

Kaikkiaan siis käsittelyjono nopeutti merkittävästi valvojien piippausta.