@vesal 2023-05-07
Kun jonoa puretaan ajetaanko kaikki jonossa olevat sisäsään yhdellä
SQL lauseella vai useilla lukiten?  Tuo yhdellä ajaminen saattaisi
lyhentää sitä käsittelyaikaa?