@dezhidki 2023-05-07
> Kun jonoa puretaan ajetaanko kaikki jonossa olevat sisäsään yhdellä SQL lauseella vai useilla lukiten? Tuo yhdellä ajaminen saattaisi lyhentää sitä käsittelyaikaa?

Totta kai kaikki muutokset ajetaan samassa transaktiossa, ja SQLAlchemy hoitaa erillisten kirjoitus- ja päivityslyselyiden yhdistämistä yhdeksi saman transaktion sisällä. 

Suurempi hidastus tulee oikeuksien levityksestä, joka vaatii oikeuslokin lukitsemista oikeusoperaatioiden käsittelyn ajaksi. Esittämäni 1,5 min käsittelyaika tulee ottaa sillä varauksella vastaan, että se on ns. "pahimman tapauksen" käsittelyaika. Testeissäni pahin tapaus oli, että 1000 hakijaa kirjataan sisään 30 sekunnissa, mikä on aika kaukana nykyisestä todellisuudesta. Lisäksi on otettava huomioon, että

* Tällä hetkellä massatenteissa sisäänkirjaamisen jälkeen hakijan pitää kävellä omalle paikalle, käynnistää tietokone ja kirjautua TIMiin. Uskon, että koko hommaan voi mennä kyllä reilusti minuutti, minkä aikana oikeudet ehtivät levittyä.
* Jos rajoite muuttuu todelliseksi ongelmaksi, voidaan pohtia erilaisia viivyttelymenetelmiä. Esimerkiksi tentin aloitussivu voidaan pitää auki kaikille, jolloin hakija pääsee aloitussivulle ilman tarvetta oikeuden levitykselle. Ajoittamalla sisäänkirjaus voidaan siten varmistaa, että hakijoiden sisäänkirjaamisen jälkeen jää 1-2 min "tyhjää" aikaa, jona käsittelyjono ehditään käsitellä loppuun.

Ja taas, tämä siis pätee siinä tapauksessa, jossa sisäänkirjausvauhti olisi $30 \frac{\text{hakijaa}}{\text{sekunti}}$.
Ainakaan valintakokeissa ei vielä ole päästy sellaiseen vauhtiin :)

En siis tässä PRssä rupea muuttamaan oikeudenlevitystä, vaan kokeillaan, miten tämä käsittelyjono auttaa. Hätätilanteessa se voidaan kytkeä nopeasti pois päältä.